feat: server services
Refactor restaurant_info to use f32 for position coordinates

Add random_bool_by_given_probability function to utils

Implement PaymentGateway and PaymentAcceptor for handling payments

Enhance PaymentGateway to manage authorized orders and network messages

Update Restaurant handler to use RequestNearbyDelivery message

Fix server main.rs to correctly parse IP address

Add StorageLogMessage enum for better storage updates management

Implement NearbyDeliveryService for fetching nearby deliveries

Create NearbyRestaurantsService for retrieving nearby restaurants

Develop OrderService to manage order lifecycle and interactions

Implement Storage actor for managing clients, restaurants, deliveries, and orders

+++ common/src/types/restaurant_info.rs
@@ -3,5 +3,5 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct RestaurantInfo {
     pub id: String,
-    pub position: (f64, f64),
+    pub position: (f32, f32),
 }
